# 5104 faster jni