refs #5104 

Expanding this PR scope with suggestion from @jfirebaugh to eliminate `getField` calls:
 - [x] nativeSetLatLng
   - from 117 to 105ms 
 - [x] nativeProjectedMetersForLatLng
  - from 32 to 30ms
 - [x] nativeLatLngForProjectedMeters
  - from 34 to 31ms
 - [x] nativeJumpTo
  - from 94 to 86ms
 - [x] nativeEaseTo
  - from 103 to 94ms
 - [x] nativeFlyTo
  - from 106 to 97ms

Out of scope for this PR:
  - Java List object manipulation in JNI (eg. addMarkers, AddPolylines,...)
  - Offline related calls