@jfirebaugh 2016-06-20
👍 

We should do this treatment to all the JNI methods that do multiple `GetField`s internally.