@zugaldia 2016-06-21
As discussed with @tobrun this is currently pointing against `master` and not urgent to include in the 4.1 release. Removing it from the milestone.